pyproject.toml
[project]
name = "cube-reconstruction"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "JustinTienken-Harder", email = "justinhardermsc@gmail.com"},
]
dependencies = [
    "requests>=2.32.3",
    "bs4>=0.0.2",
    "kociemba>=1.2.1",
    "numpy>=2.2.4",
    "scipy>=1.15.2",
    "pillow>=11.1.0",
    "pytest>=8.3.5",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
